+++ sdk/src/test/java/com/hedera/hashgraph/sdk/AccountIdTest.java
@@ -144,7 +144,7 @@ class AccountIdTest {
     }
 
     @Test
-    void fromStringWithAliasEvmAddress() {
+    void fromStringWithEvmAddress() {
         SnapshotMatcher.expect(AccountId.fromString("0.0.302a300506032b6570032100114e6abc371b82da").toString()).toMatchSnapshot();
     }
 
@@ -169,7 +169,7 @@ class AccountIdTest {
     }
 
     @Test
-    void toBytesAliasEvmAddress() {
+    void toBytesEvmAddress() {
         SnapshotMatcher.expect(Hex.toHexString(AccountId.fromString("0.0.302a300506032b6570032100114e6abc371b82da").toBytes())).toMatchSnapshot();
     }
 
@@ -205,13 +205,20 @@ class AccountIdTest {
     }
 
     @Test
-    void fromBytesAliasEvmAddress() throws InvalidProtocolBufferException {
+    void fromBytesEvmAddress() throws InvalidProtocolBufferException {
         SnapshotMatcher.expect(AccountId.fromBytes(AccountId.fromString("0.0.302a300506032b6570032100114e6abc371b82da").toBytes()).toString()).toMatchSnapshot();
     }
 
     @Test
-    void toFromProtobufAliasEvmAddress() {
+    void toFromProtobufEvmAddress() {
         var id1 = AccountId.fromString("0.0.302a300506032b6570032100114e6abc371b82da");
+        var id2 = AccountId.fromProtobuf(id1.toProtobuf());
+        assertThat(id2).isEqualTo(id1);
+    }
+
+    @Test
+    void toFromProtobufRawEvmAddress() {
+        var id1 = AccountId.fromString("302a300506032b6570032100114e6abc371b82da");
         var id2 = AccountId.fromProtobuf(id1.toProtobuf());
         assertThat(id2).isEqualTo(id1);
     }
@@ -237,5 +244,15 @@ class AccountIdTest {
         id2.virtualAddresses.add(new VirtualAddress(address2, false));
 
         assertThat(id2).isEqualTo(id1);
+    }
+
+    @Test
+    void fromEvmAddress() {
+        String evmAddress = "302a300506032b6570032100114e6abc371b82da";
+        var id = AccountId.fromEvmAddress(evmAddress, 5, 9);
+
+        assertThat(id.evmAddress.toString()).isEqualTo(evmAddress);
+        assertThat(id.shard).isEqualTo(5);
+        assertThat(id.realm).isEqualTo(9);
     }
 }